src/test/java/org/springframework/samples/petclinic/sfg/HearingInterpreterTest.java
package org.springframework.samples.petclinic.sfg;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ActiveProfiles;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.TestPropertySource;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import static org.assertj.core.api.Assertions.assertThat;

@ExtendWith(SpringExtension.class)
@ContextConfiguration(classes = { SpringTestConfiguration.class })
//@ActiveProfiles("laurel")
@TestPropertySource("/application-laurel.properties")

class HearingInterpreterTest {

    @Autowired
    HearingInterpreter hearingInterpreter;

    @Test
    void whatIheard() {
        //given

        //when
        String result = hearingInterpreter.whatIheard();

        //then
        assertThat(result).isEqualTo("Yanny");
    }
}